@@ -1,12 +1,9 @@
 package com.mzimecki.camel.spring.webservice;
 
-import javax.servlet.http.HttpServletRequest;
-
 import org.apache.camel.CamelContext;
 import org.apache.camel.Exchange;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.ExchangeBuilder;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -21,19 +18,22 @@ import com.mzimecki.camel.spring.constants.ServiceConstants;
 @RequestMapping("service")
 public class HelloService {
 	
-	@Autowired
-	private ProducerTemplate producer;
+	private final ProducerTemplate producer;
 	
-	@Autowired
-	private CamelContext camelContext;
+	private final CamelContext camelContext;
+
+	public HelloService(ProducerTemplate producer, CamelContext camelContext) {
+		this.producer = producer;
+		this.camelContext = camelContext;
+	}
 	
 	@PostMapping(value = "/hello", consumes={MediaType.TEXT_PLAIN_VALUE}, produces = {MediaType.TEXT_PLAIN_VALUE})
 	@ResponseBody
-	public ResponseEntity<?> hello(final HttpServletRequest request, @RequestBody String requestBody) {
+	public ResponseEntity<String> hello(@RequestBody String requestBody) {
 		final Exchange requestExchange = ExchangeBuilder.anExchange(camelContext).withBody(requestBody).build();
 		final Exchange responseExchange = producer.send(ServiceConstants.HELLO_SERVICE_ENDPOINT, requestExchange);
-		final String responseBody = responseExchange.getOut().getBody(String.class);
-		final int responseCode = responseExchange.getOut().getHeader(Exchange.HTTP_RESPONSE_CODE, Integer.class).intValue();
+		final String responseBody = responseExchange.getMessage().getBody(String.class);
+		final int responseCode = responseExchange.getMessage().getHeader(Exchange.HTTP_RESPONSE_CODE, Integer.class);
 		return ResponseEntity.status(responseCode).body(responseBody);
 	}
 }
